ref: move fields components to fields folder
