+++ resources/js/fields/editor-index.vue
@@ -4,7 +4,9 @@
 
 <script>
 export default {
-    props: ['resourceName', 'field'],
+    props: [
+        'resourceName', 'field'
+    ],
     computed: {
         value() {
             if (this.field.value) {
